+++ src/messages/GE.json
@@ -485,5 +485,7 @@
   "Remove from collection": "Remove from collection",
   "Invalid file format": "Looks like you tried to upload an invalid file format. Please upload a valid Postman collection file",
   "Upload collection": "Upload collection",
-  "Would you like to merge with the current collection?": "Would you like to merge with the current collection?"
+  "Would you like to merge with the current collection?": "Would you like to merge with the current collection?",
+  "Endpoints": "Endpoints",
+  "Collection": "Collection"
 }
